refactor: extract startup prompt and trust resolution helpers
