+++ crates/pi-coding-agent/src/main.rs
@@ -22,6 +22,7 @@ mod session_runtime_helpers;
 mod skills;
 mod skills_commands;
 mod slack;
+mod startup_resolution;
 mod time_utils;
 mod tool_policy_config;
 mod tools;
@@ -198,6 +199,9 @@ pub(crate) use crate::skills_commands::{
     execute_skills_trust_list_command, execute_skills_trust_revoke_command,
     execute_skills_trust_rotate_command, execute_skills_verify_command,
     render_skills_lock_write_success, render_skills_sync_drift_details, render_skills_sync_in_sync,
+};
+pub(crate) use crate::startup_resolution::{
+    ensure_non_empty_text, resolve_skill_trust_roots, resolve_system_prompt,
 };
 pub(crate) use crate::time_utils::{
     current_unix_timestamp, current_unix_timestamp_ms, is_expired_unix,
@@ -1915,66 +1919,6 @@ fn execute_channel_store_admin_command(cli: &Cli) -> Result<()> {
     }
 
     Ok(())
-}
-
-fn resolve_system_prompt(cli: &Cli) -> Result<String> {
-    let Some(path) = cli.system_prompt_file.as_ref() else {
-        return Ok(cli.system_prompt.clone());
-    };
-
-    let system_prompt = std::fs::read_to_string(path)
-        .with_context(|| format!("failed to read system prompt file {}", path.display()))?;
-
-    ensure_non_empty_text(
-        system_prompt,
-        format!("system prompt file {}", path.display()),
-    )
-}
-
-fn ensure_non_empty_text(text: String, source: String) -> Result<String> {
-    if text.trim().is_empty() {
-        bail!("{source} is empty");
-    }
-    Ok(text)
-}
-
-fn resolve_skill_trust_roots(cli: &Cli) -> Result<Vec<TrustedKey>> {
-    let has_store_mutation = !cli.skill_trust_add.is_empty()
-        || !cli.skill_trust_revoke.is_empty()
-        || !cli.skill_trust_rotate.is_empty();
-    if has_store_mutation && cli.skill_trust_root_file.is_none() {
-        bail!("--skill-trust-root-file is required when using trust lifecycle flags");
-    }
-
-    let mut roots = Vec::new();
-    for raw in &cli.skill_trust_root {
-        roots.push(parse_trusted_root_spec(raw)?);
-    }
-
-    if let Some(path) = &cli.skill_trust_root_file {
-        let mut records = load_trust_root_records(path)?;
-        if has_store_mutation {
-            let report = apply_trust_root_mutations(&mut records, cli)?;
-            save_trust_root_records(path, &records)?;
-            println!(
-                "skill trust store update: added={} updated={} revoked={} rotated={}",
-                report.added, report.updated, report.revoked, report.rotated
-            );
-        }
-
-        let now_unix = current_unix_timestamp();
-        for item in records {
-            if item.revoked || is_expired_unix(item.expires_unix, now_unix) {
-                continue;
-            }
-            roots.push(TrustedKey {
-                id: item.id,
-                public_key: item.public_key,
-            });
-        }
-    }
-
-    Ok(roots)
 }
 
 pub(crate) fn write_text_atomic(path: &Path, content: &str) -> Result<()> {
